Changed the iteration indices and any datatype type associated with ints or unsigned ints to long long ints and unsigned long long ints. This enables the data structures to deal with very large array sizes and numeric values.All the libraries were modified accordingly.

+++ Sort/stack.h
@@ -57,8 +57,8 @@ typedef void    (*StackDestroyFn)(void *);
 typedef struct
 {
     void                **A;
-    int                 n;
-    int                 size;
+    long long int       n;
+    long long int       size;
     StackCompareFn      cmp;
     StackPrintFn        print;
     StackDestroyFn      destroy;
